app/requirements.txt
fastapi<=0.68.2
uvicorn<=0.15.0
sqlalchemy<=1.4.41
psycopg2-binary<=2.9.6
python-jose<=3.3.0
passlib<=1.7.4